requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pydub==0.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
boto3==1.34.10
Pillow==10.1.0
